refactor: refatorando componente items-page @MarquesEmilly @nicofercavv-dev
